Refactored Names, Directories, Modules
Also added singleton methods to gameengine and assetmanager.
