+++ Core/Camera.js
@@ -1,0 +1,3 @@
+import { Entity } from "../Entities/Entities.js";
+
+class Camera extends Entity {}
