--- glc.js
@@ -1,16 +1,0 @@
-class GameLogicController extends Entity {
-  constructor() {
-    super();
-    this.entityOrder = -1;
-    this.state = 0;
-  }
-
-  update() {
-    if (this.state == 0) {
-      this.state = 1;
-      gameEngine.addEntity(new Player());
-      gameEngine.addEntity(new Background());
-      gameEngine.addEntity(new GameMap());
-    }
-  }
-}
