+++ index.html
@@ -5,16 +5,20 @@
     <meta charset="utf-8">
     <title>Game Project Title Goes Here</title>
   
-    <script type="text/javascript" src="./assetmanager.js"></script>
-    <script type="text/javascript" src="./timer.js"></script>
-    <script type="text/javascript" src="./gameengine.js"></script>
-    <script type="text/javascript" src="./util.js"></script>
-    <script type="text/javascript" src="./entitys.js"></script>
-    <script type="text/javascript" src="./player.js"></script>
-    <script type="text/javascript" src="./background.js"></script>
-    <script type="text/javascript" src="./camera.js"></script>
-    <script type="text/javascript" src="./glc.js"></script>
-    <script type="text/javascript" src="./main.js"></script>
+    <script type="module" src="./Core/AssetManager.js"></script>
+    <script type="module" src="./Core/GameEngine.js"></script>
+    <script type="module" src="./Core/GameLogicController.js"></script>
+    <script type="module" src="./Core/Camera.js"></script>
+    
+    <script type="module" src="./Entities/Player.js"></script>
+    <script type="module" src="./Entities/Entities.js"></script>
+    <script type="module" src="./Entities/Background.js"></script>
+    
+    <script type="module" src="./Utils/Util.js"></script>
+    <script type="module" src="./Utils/Timer.js"></script>
+    
+    <script type="module" src="./main.js"></script>
+
 </head>
 <body style="background-color: black;">
     <canvas id="gameWorld" width="1024" height="768" style="background: white; border: 1px solid black; display: block; margin: auto; padding: 0; position: absolute;
